+++ drivers/mm/mm_drv_intel_adsp_tlb.c
@@ -27,7 +27,7 @@
 #include <zephyr/spinlock.h>
 #include <zephyr/sys/__assert.h>
 #include <zephyr/sys/check.h>
-#include <zephyr/sys/mem_manage.h>
+#include <zephyr/kernel/mm.h>
 #include <zephyr/sys/util.h>
 #include <zephyr/debug/sparse.h>
 #include <zephyr/cache.h>
