+++ drivers/pcie/host/msi.c
@@ -38,7 +38,7 @@ static uint32_t pcie_msi_base(pcie_bdf_t bdf, bool *msi)
 
 #ifdef CONFIG_PCIE_MSI_MULTI_VECTOR
 
-#include <zephyr/sys/mem_manage.h>
+#include <zephyr/kernel/mm.h>
 
 __weak uint8_t arch_pcie_msi_vectors_allocate(unsigned int priority,
 					      msi_vector_t *vectors,
