+++ include/zephyr/arch/x86/ia32/linker.ld
@@ -40,7 +40,7 @@
 #include <zephyr/linker/linker-defs.h>
 #include <offsets.h>
 #include <zephyr/sys/util.h>
-#include <zephyr/sys/mem_manage.h>
+#include <zephyr/kernel/mm.h>
 
 #include <zephyr/linker/linker-tool.h>
