+++ tests/kernel/fatal/exception/src/main.c
@@ -13,7 +13,7 @@
 #include <assert.h>
 
 #if defined(CONFIG_USERSPACE)
-#include <zephyr/sys/mem_manage.h>
+#include <zephyr/kernel/mm.h>
 #include <zephyr/internal/syscall_handler.h>
 #include "test_syscalls.h"
 #endif
